requirements.txt
gitpython
psutil